execute1: Improve timing for execute bypass tag
The tags for the bypass data paths back to decode2 don't really need
to depend on the stall/busy inputs or on whether an exception might be
generated, since the bypass values won't be used until the instruction
gets executed.  Therefore, this simplifies the expressions for
bypass_data.tag.valid and bypass_cr_data.tag.valid.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>