dcache: Implement dcbz to non-cacheable memory properly
A dcbz operation to memory that is mapped as non-cacheable in the page
tables doesn't cause an alignment interrupt, but neither was it
implemented properly in the dcache.  It does do 8 writes to memory but
it also creates a zero-filled line in the cache.

This fixes it so that dcbz to memory mapped non-cacheable doesn't
write the cache tag or set any line valid.  We now have r1.reloading
which is 1 only in RELOAD_WAIT_ACK state, but only if the memory is
cacheable and therefore the cache should be updated (i.e. it is zero
in RELOAD_WAIT_ACK state if we are doing a non-cacheable dcbz).
We can now also remove the code in loadstore1 that checks for
non-cacheable dcbz, which only triggered when doing dcbz in real mode
to an address in the Cxxxxxxx range.

Also remove some unused variables and signals.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>