dcache: Rework forwarding and same-page logic
This gets rid of some largish comparators in the dcache_request
process by matching index and way that hit in the cache tags instead
of comparing tag values.  That is, some tag comparisons can be
replaced by seeing if both tags hit in the same cache way.

When reloading a cache line, we now set it valid at the beginning of
the reload, so that we get hits to compare.  While the reload is still
occurring, accesses to doublewords that haven't yet been read are
indicated with req_is_hit = 0 and req_hit_reload = 1 (i.e. are
considered to be a miss, at least for now).

For the comparison of whether a subsequent access is to the same page
as stores already being performed, in virtual mode (TLB being used) we
now compare the way and index of the hit in the TLB, and in real mode
we compare the effective address.  If any new entry has been loaded
into the TLB since the access we're comparing against, then it is
considered to be a different page.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>