core: Simplify operand presentation for hash instructions
This removes the cases in the decode stages which allowed the C
register address to come from the RB field for the hash instructions
(hashst[p], hashchk[p]), and generated a negative immediate value for
the B operand.  The motivation is to simpify the logic for the C
register address.  Instead the unusual construction of the address for
the hash instructions is handled in the loadstore1_in process, and the
hash computation uses the A and B operands rather than A and C.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>