logical: Use sub_select rather than insn_type to select logical op
Also select the RS passthrough in the logical unit by default for
mfspr, which is needed for the no-op SPRs and the no-op behaviour
of privileged mfspr to unimplemented SPRs.  For slow SPRs the RS
behaviour gets passed through from execute1 to execute2 and
replaced by the correct result in execute2's result mux.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>