execute1: Fix no-op behaviour of reading undefined SPRs
In privileged mode, mfspr from an undefined or unimplemented SPR
number should be a no-op, which is implemented here by writing back
the same value that the destination register previously had.  However,
we ended up writing back 0 because ex1.res2_sel was not set correctly.
To fix this, set res2_sel to 10 in the undefined SPR case.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>