FPU: Renormalize frsp operand if denormalized
This arranges for the frsp operand to be renormalized if necessary.
Without this, we can incorrectly get X set to 1 for denormalized
operands, and hence the rounding may be done incorrectly.  To make
things clearer, we now have an explicit flag indicating when the B
operand needs to be in normalized form.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>